Pass a dict, instead of None, to modules if config is None
If a Synapse module's config block were empty in YAML, thus being translated
to a Nonetype in Python, then some modules could fail as they expect a dict
to be passed to their parse_config method.

This PR ensures that if the user does provide a None config block, such as
what the default oidc config provides, then that None is not passed to the
module. An empty dict is passed instead.

This code assumes that no existing modules are relying on receiving a None
config block, but I'd really hope that they aren't.