Use previous depth if overriding prev_events to insert into history
TODO: Is the assumption of anytime we pass in prev_event_ids, we use same depth
good enough? What corner cases are there? I see that we also pass in prev_event_ids this in
synapse/handlers/room_member.py so need to make sure that still work as expected