Fix tests to check the shared_rooms endpoints multiple times
Beforehand, with the cache added, multiple calls would give back the same
results, even if the users started sharing a new room in-between those calls.

Removing the cache should fix the problem, but it's a good idea to add tests
to cache the mistake in the future; in case we do decide to add a cache again.