Modify `get_interested_parties` and `get_interested_remotes` to query PresenceRouter
This commit asks the PresenceRouter for any users - in addition to those
from users that they share a room with - that should receive the given
presence updates.

These methods are called when routing new presence updates around as
they come in.

* `get_interested_parties` is called when figuring out which local and
remote users to send presence to. For local users, their sync streams
will be woken up.
* `get_interested_remotes` is specifically for figuring out which remote
user(s) a given presence update needs to go to.