Add load-balancing support; support sharding media_repo
This required some changes to the way we handle our nginx config. It's
now a dictionary instead of just a str. This was because endpoint
mapping blocks would get duplicated when we configured multiple
instances of the same worker. Now that the nginx config is organised by
a map of endpoints to upstreams, those endpoints de-duplicate
automatically.

Speaking of upstreams, we now handle creating upstreams which load
balance between processes. To route traffic to an upstream, you simply
need to proxy_pass to its name. We use a separate dict to keep track of
upstreams and where they point to. These upstreams do round-robin LBing.

Note that we don't really want other load-balancing types for our tests. For instance,
balancing by request IP makes no sense as the tests will be sending requests
all the from the same IP.

What we really want is for the load-balancing to be deterministic, which
round-robin is (as long as the tests remain constant themselves).