Add a new key to presence_set_state replication request, force_notify
If force_notify is True, we will create a new stream ID for this
presence update, even if it matches the user's previous presence update
state. Creating a new stream ID will notify clients, hence
'force_notify'.

We're resending the current state of the user who's getting full presence
in order to force the presence stream ID to advance 1 - as we compare
stream IDs when checking if a device (who's presencing a presence stream
ID as part of their sync token) has already received full online
presence as part of a previous sync.

If we didn't do so, it's possible for a syncing device to receive full
online presence multiple times if the presence stream doesn't advance
otherwise.