self -> test_case. Fix worker-based test
It turned out we needed to advance the reactor in order to successfully
write the body of the replication request from the worker to the main
process.

We do this so that we can test calling send_local_online_presence_to on
a worker, while a client is syncing via the main process.