Organise existing documentation into a tree structure.
SUMMARY.md defines the sidebar of the website. We organise existing documentation
into a logical structure. We also add a couple of pages which make use of mdbook's
file include feature to include existing documentation that typically lives
outside of the docs/ directory. We also use it to combine the sample config files
into a single page with surrounding contextual info.