Properly prevent rc versions from causing a deploy
We were checking a version number derived from the branch for whether it
included "rc" - which didn't make sense, branches never have "rc" in the
name. Instead, we use Synapse's __version__ string to perform the rc
check.

There's some potential to using tags instead here, but this works for
now...