Remove the unused public_room_list_stream
Modifications to this stream were only possible on the main process, which
had the knock-on effect of other tasks, mainly creating a room, only being
possible on the main process as well.

Instead of rearchitecting this stream to allow multiple writers though, we
can actually just remove it entirely - it's not used.