Include room avatars in email notifications
Judging by the template, this was intended ages ago, but we never
actually passed an avatar URL to the template.

I'm not sure what the best way to test this is. I've written a unit test
of sorts, but I'd like to manually verify it by actually seeing an
email. Maybe I should setup my own installation of synapse?

Closes #1546.