Always add local users to the user directory
This means the config option `search_all_users` ends up controlling the
search query only, and not the data we store. The cost of doing so is an
extra row in the `user_directory` and `user_directory_search` tables for
each local user in zero rooms. I think the cost of this will be
marginal (since they'll already have entries in `users` and `profiles`
anyway).

As a small upside, a homeserver whose directory was built with this
change can toggle `search_all_users` without having to rebuild their
directory.