Fix iteration over row results
We were originally iterating over 'txn', which can work, but as 'txn' was being used to run a
simple_delete operation inside the loop, it was actually being redefined mid-loop.

Thus, this caused the background job that was intended to remove old pushers that were
associated with an unlinked email address to fail if there were any to remove.